Fixes tests and adds missing file
